splitted sdcdevice subscriptionmngr.py into two files
